requirements.txt
langchain
spacy
google-generativeai
streamlit
python-docx 
PyPDF2 
pdfplumber 
spacy  
pandas 
nltk 
scikit-learn
python-dotenv
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl 
